requirements.txt
#medical imaging
SimpleITK==2.4.0
nibabel==5.3.2
monai==1.4.0
foundation-cancer-image-biomarker==1.0.0
torchio==0.20.19


nlpaug==1.1.11
tensorboardX==2.6.2.2
betacal==1.1.0

#openai clip
ftfy==6.3.1
regex==2023.12.25
tqdm==4.66.2
git+https://github.com/openai/CLIP.git
